requirements.txt
flask==3.0.0
groq==0.4.2
reportlab==4.0.7
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
